+++ lib/deltaq/src/DeltaQ/Methods.hs
@@ -17,6 +17,10 @@ module DeltaQ.Methods
     , SlackOrHazard (..)
     , isSlack
     , isHazard
+    
+    -- * Timeouts
+    , retryOverlapN
+    , retryOverlap
     ) where
 
 import DeltaQ.Class
@@ -24,6 +28,10 @@ import DeltaQ.Class
     , Eventually (..)
     , Outcome (..)
     , eventually
+    )
+import DeltaQ.PiecewisePolynomial
+    ( DQ
+    , timeout
     )
 
 {-----------------------------------------------------------------------------
@@ -83,3 +91,45 @@ meetsRequirement o (t,p)
     p' = o `successWithin` t
 
     err = error "distanceToReference: inconsistency"
+
+{-----------------------------------------------------------------------------
+    Methods
+    Timeouts
+------------------------------------------------------------------------------}
+
+-- | Retry an outcome @n@ times, take the first instance that succeeds.
+--
+-- More specifically, @retryOverlapN n dt o@ models the completion times
+-- of a race of @n@ independent copies of the outcome @o@ against each other,
+-- succeeding when any of the copies first succeeds.
+--
+-- Copy number @0 <= j < n@ is started after waiting a time @j * dt@.
+--
+-- If no copy succeeds within total time @n * dt@,
+-- the whole outcome fails.
+retryOverlapN :: Int -> Duration DQ -> DQ -> DQ
+retryOverlapN n dt = retryOverlap (replicate n dt)
+
+-- | Retry an outcome multiple times, take the first instance that succeeds.
+--
+-- More specifically, @retryOverlap dts o@ models the completion times
+-- of a race of @length dts@ independent copies of the outcome @o@
+-- against each other, succeeding when any of the copies first succeeds.
+--
+-- Each copy is started after a timeout
+-- in the event that the previous copies have not finished yet;
+-- the list @dts@ gives the successive timeouts.
+--
+-- If no copy succeeds within total time @sum dts@,
+-- the whole outcome fails.
+retryOverlap :: [Duration DQ] -> DQ -> DQ
+retryOverlap dts0 o = go dts0 o
+  where
+    go [] _ = never
+    go (dt:dts) race =
+        choice p before (wait dt .>>. retry)
+      where
+        (before, p, after) = timeout dt race
+
+        -- add another copy of `o` to the race
+        retry = go dts (after .\/. o)